Finish grouping by encoding-specificity: observables and Kernel alias
select_observable is LCU-specific (built from the encoding's signed
Pauli terms), so it moves to pauli_lcu next to the rest of the LCU
surface; PauliLCU's protocol hook now calls it locally instead of
deferred-importing qubitization. The generic _bit_projector helper
moves to common_kernels, where both observables' modules share it.
qubitization keeps only the geometry-only reflection_observable.

The Kernel alias's canonical home is block_encoding (it describes the
protocol's factory return type, nothing LCU); pauli_lcu, qubitization,
and qsvt now type-check against that single definition instead of a
duplicate in pauli_lcu.

Package-level exports unchanged.
